requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
Babel==2.8.0
beautifulsoup4==4.8.2
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
Django==3.0.1
django-background-tasks==1.2.5
django-celery==3.3.1
django-ckeditor==6.0.0
django-compat==1.0.15
django-crispy-forms==1.9.2
django-filter==2.4.0
django-js-asset==1.2.2
django-modelcluster==5.1
django-phonenumber-field==5.0.0
django-taggit==1.3.0
django-tempus-dominus==5.1.2.13
django-treebeard==4.3.1
django-twilio==0.13.0.post1
djangorestframework==3.12.1
docopt==0.4.0
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
Jinja2==2.11.2
kombu==3.0.37
l18n==2018.5
mailchimp==2.0.9
MarkupSafe==1.1.1
openpyxl==3.0.5
paystackapi==2.0.0
phonenumbers==8.12.11
Pillow==7.2.0
PyJWT==1.7.1
python-http-client==3.3.1
pytz==2020.1
requests==2.24.0
sendgrid==6.4.7
six==1.15.0
soupsieve==2.0.1
South==1.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
tablib==2.0.0
twilio==6.45.4
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10.2
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
